PythonCode/requirements.txt
colorama
pytest
python-dotenv